+++ Plugins/UVOCBot.Plugins.Planetside/Extensions/IServiceCollectionExtensions.cs
@@ -2,9 +2,11 @@
 using DbgCensus.EventStream.EventHandlers.Extensions;
 using DbgCensus.Rest;
 using DbgCensus.Rest.Extensions;
+using DbgCensus.Rest.Objects;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Remora.Commands.Extensions;
+using Remora.Discord.Commands.Extensions;
 using UVOCBot.Plugins.Planetside;
 using UVOCBot.Plugins.Planetside.Abstractions.Services;
 using UVOCBot.Plugins.Planetside.CensusEventHandlers;
@@ -19,8 +21,9 @@ public static class IServiceCollectionExtensions
     public static IServiceCollection AddPlanetsidePlugin(this IServiceCollection services, IConfiguration config)
     {
         services.Configure<PlanetsidePluginOptions>(config.GetSection(nameof(PlanetsidePluginOptions)));
-        services.Configure<CensusQueryOptions>(config.GetSection(nameof(CensusQueryOptions)));
         services.Configure<EventStreamOptions>(config.GetSection(nameof(EventStreamOptions)));
+        services.Configure<CensusQueryOptions>(config.GetSection(nameof(CensusQueryOptions)));
+        services.Configure<CensusQueryOptions>(o => o.LanguageCode = CensusLanguage.English);
 
         services.AddHttpClient();
         services.AddSingleton<IPopulationService, HonuPopulationService>();
@@ -35,10 +38,12 @@ public static class IServiceCollectionExtensions
         services.AddPayloadHandler<MetagameEventResponder>();
 
         services.AddCommandTree()
+                .WithCommandGroup<CharacterCommands>()
                 .WithCommandGroup<OtherCommands>()
                 .WithCommandGroup<OutfitTrackingCommands>()
                 .WithCommandGroup<WorldCommands>()
-                .Finish();
+                .Finish()
+                .AddAutocompleteProvider<CharacterNameAutocompleteProvider>();
 
         services.AddHostedService<EventStreamWorker>();
         services.AddHostedService<CensusStateWorker>();