+++ Plugins/UVOCBot.Plugins.Planetside/Services/CensusApiService.cs
@@ -53,18 +53,32 @@ public class CensusApiService : ICensusApiService, IDisposable
             .OnCollection("outfit")
             .Where("alias_lower", SearchModifier.Equals, tag.ToLower());
 
-        return await GetAsync<Outfit?>(query, ct).ConfigureAwait(false);
+        return await GetAsync<Outfit>(query, ct).ConfigureAwait(false);
+    }
+
+    public virtual async Task<Result<List<Outfit>>> GetOutfitsAsync(IEnumerable<ulong> outfitIDs, CancellationToken ct = default)
+    {
+        List<ulong> outfitIDList = outfitIDs.ToList();
+        if (outfitIDList.Count == 0)
+            return new List<Outfit>();
+
+        IQueryBuilder query = _queryService.CreateQuery()
+            .OnCollection("outfit")
+            .WhereAll("outfit_id", SearchModifier.Equals, outfitIDList);
+
+        return await GetListAsync<Outfit>(query, ct);
     }
 
     /// <inheritdoc />
-    /// <remarks>This query is cached.</remarks>
     public virtual async Task<Result<Outfit?>> GetOutfitAsync(ulong id, CancellationToken ct = default)
     {
-        IQueryBuilder query = _queryService.CreateQuery()
-            .OnCollection("outfit")
-            .Where("outfit_id", SearchModifier.Equals, id);
+        Result<List<Outfit>> outfitsResult = await GetOutfitsAsync(new[] { id }, ct);
+        if (!outfitsResult.IsSuccess)
+            return Result<Outfit?>.FromError(outfitsResult);
 
-        return await GetAsync<Outfit?>(query, ct).ConfigureAwait(false);
+        return outfitsResult.Entity.Count == 0
+            ? default
+            : outfitsResult.Entity[0];
     }
 
     /// <inheritdoc/>